Replace degree with section symbol in law summary
For ticket 12602, the degree symbol is replacing the seciton symbol. Our
processor was not replacing that symbol for the summary section. By
adding a replace all, this solves the problem on our end.

This was also an issue with the daybreak, which means they must also have a
data problem as well.